requirements.txt
anyio==4.3.0
trio==0.25.0
httpx==0.27.0
pydantic==1.10.0
pytest-httpx==0.30.0
pytest-asyncio==0.23.6
